Api cahnges

+++ Zbang.Cloudents.Mvc4WebRole/App_Start/BundleConfig.cs
@@ -231,11 +231,10 @@ namespace Zbang.Cloudents.Mvc4WebRole
                 {
                     var relativePath = filePath.Replace(appRoot, string.Empty);
                     cdnFilePath = Path.Combine(cdnRoot, relativePath);
-                    // we want to replace the file if its exits
-                    //if (File.Exists(Path.Combine(cdnRoot, relativePath)))
-                    //{
-                    //    continue;
-                    //}
+                    if (File.Exists(Path.Combine(cdnRoot, relativePath)))
+                    {
+                        continue;
+                    }
                     var directory = Path.GetDirectoryName(cdnFilePath);
                     if (directory != null) Directory.CreateDirectory(directory);
                     File.Copy(filePath, Path.Combine(cdnRoot, relativePath), true);
